requirements_live.txt
alpaca-trade-api==3.3.2
pandas-market-calendars==4.4.1
schedule==1.2.2
polygon-api-client==1.14.0
